+++ src/main/java/com/aetherteam/aether/mixin/mixins/common/PowderSnowBlockMixin.java
@@ -1,6 +1,6 @@
 package com.aetherteam.aether.mixin.mixins.common;
 
-import com.aetherteam.aether.accessories.compat.AccessoryEffectBridge;
+import com.aetherteam.aether.accessories.effect.AccessoryEffectBridge;
 import com.llamalad7.mixinextras.injector.ModifyReturnValue;
 import net.fabricmc.fabric.api.util.TriState;
 import net.minecraft.world.entity.Entity;