+++ src/main/java/com/aetherteam/aether/mixin/mixins/common/compat/twilight/LifedrainScepterItemMixin.java
@@ -1,6 +1,6 @@
 package com.aetherteam.aether.mixin.mixins.common.compat.twilight;
 
-import com.aetherteam.aether.accessories.compat.AccessoryEffectBridge;
+import com.aetherteam.aether.accessories.effect.AccessoryEffectBridge;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;